Moving station location data into libmetoceanviewer
