+++ MetOceanViewer_GUI/movResource.qrc
@@ -1,5 +1,5 @@
 <RCC>
-    <qresource prefix="/rsc" lang="Generic Resources">
+    <qresource lang="Generic Resources" prefix="/rsc">
         <file>img/mm_20_darkblue.png</file>
         <file>img/mm_20_darkgreen.png</file>
         <file>img/mm_20_darkorange.png</file>
@@ -17,7 +17,7 @@
         <file>img/arrow_down.png</file>
         <file>img/arrow_up.png</file>
     </qresource>
-    <qresource prefix="/qml" lang="QML">
+    <qresource lang="QML" prefix="/qml">
         <file>qml/MovMapItem.qml</file>
         <file>qml/MapViewer.qml</file>
         <file>qml/InfoWindow.qml</file>
@@ -25,11 +25,5 @@
         <file>qml/EsriMapViewer.qml</file>
         <file>qml/MapboxMapViewer.qml</file>
         <file>qml/OsmMapViewer.qml</file>
-    </qresource>
-    <qresource prefix="/stations" lang="csv">
-        <file>data/noaa_stations.csv</file>
-        <file>data/usgs_stations.csv</file>
-        <file>data/xtide_stations.csv</file>
-        <file>data/ndbc_stations.csv</file>
     </qresource>
 </RCC>
