+++ MetOceanViewer_GUI/src/mapfunctions.h
@@ -29,13 +29,10 @@
 class MapFunctions : public QObject {
   Q_OBJECT
  public:
-  enum MarkerType { NOAA, USGS, XTIDE, NDBC };
 
   enum MapSource { ESRI, MapBox, OSM };
 
   explicit MapFunctions(QObject *parent = nullptr);
-
-  QVector<Station> readMarkers(MapFunctions::MarkerType markerType);
 
   int refreshMarkers(StationModel *model, QQuickWidget *map,
                      QVector<Station> &locations, bool filter = true,
