+++ libraries/libmetoceanviewer/data/ndbc/getNdbcLocs.sh
