+++ libraries/libmetoceanviewer/data/ndbc_stations.csv
