+++ libraries/libmetoceanviewer/data/noaa/stations.txt
