+++ libraries/libmetoceanviewer/resource_files.qrc
@@ -2,4 +2,10 @@
     <qresource prefix="/rsc">
         <file>harmonics.tcd</file>
     </qresource>
+    <qresource prefix="/stations">
+        <file>data/ndbc_stations.csv</file>
+        <file>data/noaa_stations.csv</file>
+        <file>data/usgs_stations.csv</file>
+        <file>data/xtide_stations.csv</file>
+    </qresource>
 </RCC>
