+++ libraries/libmetoceanviewer/station.h
@@ -25,6 +25,7 @@
 
 class Station {
  public:
+
   Station();
   Station(QGeoCoordinate coordinate, QString id, QString name,
           double measured = 0.0, double modeled = 0.0, int category = 0,
@@ -71,6 +72,7 @@ class Station {
   void setActive(bool active);
 
  private:
+
   QGeoCoordinate m_coordinate;
   QString m_name;
   QString m_id;
